requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==25.2
django-extensions==4.1
python-decouple==3.8
Pillow==12.0.0
reportlab==4.4.7
openpyxl==3.1.5
PyYAML==6.0.3
gunicorn==20.1.0
python-dotenv==0.21.0
